refactor: integrate code review comments
Related to issue #256
