--- annotations/component-annotations/src/it/component-issue-254-exposeService-not-overrided/pom.xml
@@ -1,64 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-	<modelVersion>4.0.0</modelVersion>
-	<parent>
-		<groupId>org.springframework.boot</groupId>
-		<artifactId>spring-boot-starter-parent</artifactId>
-		<version>2.1.5.RELEASE</version>
-		<relativePath/> <!-- lookup parent from repository -->
-	</parent>
-	<groupId>io.dekorate</groupId>
-	<artifactId>component-issue-254-exposeService-not-overrided</artifactId>
-	<version>0.0.1-SNAPSHOT</version>
-  <name>Dekorate :: Annotation :: Component :: Integration Test for #254</name>
-	<description>Regression test for issue #254. Ensure that exposeService=true works when combining annotations with properties.</description>
-
-	<dependencies>
-        <dependency>
-            <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-starter-web</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>io.dekorate</groupId>
-            <artifactId>kubernetes-spring-starter</artifactId>
-            <version>@project.version@</version>
-        </dependency>
-        <dependency>
-          <groupId>io.dekorate</groupId>
-          <artifactId>component-annotations</artifactId>
-          <version>@project.version@</version>
-        </dependency>
-
-        <!-- Testing -->
-        <dependency>
-          <groupId>org.junit.jupiter</groupId>
-          <artifactId>junit-jupiter</artifactId>
-          <version>@version.junit-jupiter@</version>
-          <scope>test</scope>
-        </dependency>
-  </dependencies>
-
-  <build>
-    <plugins>
-      <plugin>
-        <groupId>org.apache.maven.plugins</groupId>
-        <artifactId>maven-surefire-plugin</artifactId>
-        <version>@version.maven-surefire-plugin@</version>
-        <inherited>true</inherited>
-        <configuration>
-          <useSystemClassLoader>false</useSystemClassLoader>
-        </configuration>
-      </plugin>
-      <plugin>
-        <groupId>org.apache.maven.plugins</groupId>
-        <artifactId>maven-compiler-plugin</artifactId>
-        <version>@version.maven-compiler-plugin@</version>
-        <configuration>
-          <source>@java.source@</source>
-          <target>@java.target@</target>
-        </configuration>
-      </plugin>
-    </plugins>
-  </build>
-</project>
