--- annotations/component-annotations/src/it/component-issue-65-avoid-Env-duplication/pom.xml
@@ -1,90 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-Copyright 2018 The original authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-
--->
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-
-  <groupId>io.dekorate</groupId>
-  <artifactId>component-issue-65-avoid-Env-duplication</artifactId>
-  <version>1.0-SNAPSHOT</version>
-  <packaging>jar</packaging>
-  <name>Dekorate :: Annotation :: Component :: Integration Test for #65</name>
-  <description>Env annotation appears twice within yaml generated</description>
-
-  <properties>
-    <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-    <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-    <java.version>1.8</java.version>
-
-    <version.spring-boot>2.1.0.RELEASE</version.spring-boot>
-    <version.junit-jupiter>5.1.0</version.junit-jupiter>
-    <version.junit-platform>1.1.0</version.junit-platform>
-
-    <version.maven-surefire-plugin>3.0.0-M3</version.maven-surefire-plugin>
-  </properties>
-
-  <dependencies>
-    <dependency>
-      <groupId>io.dekorate</groupId>
-      <artifactId>kubernetes-annotations</artifactId>
-      <version>@project.version@</version>
-    </dependency>
-    <dependency>
-      <groupId>io.dekorate</groupId>
-      <artifactId>component-annotations</artifactId>
-      <version>@project.version@</version>
-    </dependency>
-    <!-- Testing -->
-    <dependency>
-      <groupId>org.junit.jupiter</groupId>
-      <artifactId>junit-jupiter-api</artifactId>
-      <version>${version.junit-jupiter}</version>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
-      <groupId>org.junit.jupiter</groupId>
-      <artifactId>junit-jupiter-engine</artifactId>
-      <version>${version.junit-jupiter}</version>
-      <scope>test</scope>
-    </dependency>
-  </dependencies>
-
-  <build>
-    <plugins>
-      <plugin>
-        <groupId>org.apache.maven.plugins</groupId>
-        <artifactId>maven-surefire-plugin</artifactId>
-        <version>${version.maven-surefire-plugin}</version>
-        <inherited>true</inherited>
-        <configuration>
-          <useSystemClassLoader>false</useSystemClassLoader>
-        </configuration>
-      </plugin>
-      <plugin>
-        <groupId>org.apache.maven.plugins</groupId>
-        <artifactId>maven-compiler-plugin</artifactId>
-        <version>@version.maven-compiler-plugin@</version>
-        <configuration>
-          <source>@java.source@</source>
-          <target>@java.target@</target>
-        </configuration>
-      </plugin>
-    </plugins>
-  </build>
-</project>
