+++ annotations/component-annotations/src/it/feat-229-override-annotationbased-config/src/main/java/io/dekorate/examples/component/Controller.java
