+++ annotations/component-annotations/src/it/feat-229-override-annotationbased-config/src/main/resources/application.yml
