+++ annotations/component-annotations/src/it/feat-275-provide-port/src/main/java/io/dekorate/issue275/DemoApplication.java
