+++ annotations/component-annotations/src/it/feat-275-provide-port/src/main/resources/application.properties
