+++ annotations/component-annotations/src/it/feat-52-add-build-spec/src/main/java/io/dekorate/examples/component/Main.java
