+++ annotations/component-annotations/src/it/feat-71-make-link-standalone/src/main/java/io/dekorate/examples/component/Main.java
