+++ annotations/component-annotations/src/it/feat-71-make-link-standalone/src/test/java/io/dekorate/examples/component/Issue71Test.java
