+++ annotations/component-annotations/src/it/issue-188-json-yml-generation/pom.xml
