+++ annotations/component-annotations/src/it/issue-188-json-yml-generation/src/main/java/io/dekorate/examples/component/Main.java
