+++ annotations/component-annotations/src/it/issue-188-json-yml-generation/src/test/java/io/dekorate/examples/component/Issue188Test.java
