+++ annotations/component-annotations/src/it/issue-193-k8s-annotations-population/src/main/java/io/dekorate/examples/component/Main.java
