+++ annotations/component-annotations/src/it/issue-193-k8s-annotations-population/src/test/java/io/dekorate/examples/component/Issue193Test.java
