+++ annotations/component-annotations/src/it/issue-254-exposeService-not-overrided/src/main/java/io/dekorate/issue254/Controller.java
