+++ annotations/component-annotations/src/it/issue-254-exposeService-not-overrided/src/main/resources/application.properties
