+++ annotations/component-annotations/src/it/issue-276-componentName-not-overrided/src/main/java/io/dekorate/issue276/Controller.java
