+++ annotations/component-annotations/src/it/issue-276-componentName-not-overrided/src/main/resources/application.properties
