+++ annotations/component-annotations/src/it/issue-276-componentName-not-overrided/src/test/java/io/dekorate/issue276/Issue276Test.java
