+++ annotations/component-annotations/src/it/issue-65-avoid-Env-duplication/src/main/java/io/dekorate/examples/component/Main.java
