+++ annotations/component-annotations/src/it/issue-65-avoid-Env-duplication/src/test/java/io/dekorate/examples/component/Issue65Test.java
