--- examples/component-example/src/main/resources/application.yml
@@ -1,10 +1,0 @@
-dekorate:
-  link:
-    name: "hello-world"
-    componentName: "target"
-    envs:
-      - name: "key1"
-        value: "val1"
-  component:
-    buildType: "docker"
-
